Drop superflous method lookup during inlining
While refactoring inlining, I noticed that the bail path (using :invoke,
rather than inlining) would repeat the method lookup we had just
completed (or that had been forwarded to us from inference). As far
as I can tell, this is unnecessary and a vestage from many years
ago when :invoke nodes were introduced at a point where the method
match was not available and the API has simply been moved around since.
As best I can tell, the practical effect of the current call is
simply to create a MethodInstance specialized to the compilation
signature (we do previously create a MethodInstance for the exact
signature we want to inline, but obviously that's slightly different).

Am I missing anything?